finished tests, adding some services, all tests pass so far

+++ backend/src/main/java/com/example/backend/repository/MajorGroupRepository.java
@@ -4,5 +4,5 @@ import com.example.backend.model.MajorGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MajorGroupRepository extends JpaRepository<MajorGroup, Integer> {
+public interface MajorGroupRepository extends JpaRepository<MajorGroup, Long> {
 }
